requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.21
Werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
requests==2.31.0
